POC integrated client/server test w/ PhantomJS
So far we've only proved that we can load a HTML file from the server
with crypton.js, and that the resulting page does indeed have a crypton
javascript object.

In general this gives us a way to drive client side code from within
node/mocha, with a unified console.log.

Next we can add tests to make the client interact w/ the server.