Prove concepts w/ server / client integration tests
We now have the ability to run a test with spawns a browser, points the
browser to a given JS test file, waits for the browser to signal that
it's finished running the test sequence (by making a GET call to the
server), and then evaluating a function in browser context to collect
results.

This gives us a framework for end-to-end functional/integration testing.

Added dependency of Q for deferred events, and making the
in-browser-context function wait on the signal from the client.